Minor fix for locale list

+++ app/header.php
@@ -64,8 +64,8 @@ if (!_isAllowed($request)) {
 	throw new SystemException(_translate("Sorry") . " {$_SESSION['userName']}. " . _translate('You do not have permission to access this page.'), 401);
 }
 $countryCode = $arr['default_phone_prefix'] ?? '';
-$minNumberOfDigits = $arr['min_phone_length'] ?? 0;
-$maxNumberOfDigits = $arr['max_phone_length'] ?? 15;
+$minNumberOfDigits = (int) ($arr['min_phone_length'] ?? 0);
+$maxNumberOfDigits = (int) ($arr['max_phone_length'] ?? 15);
 
 $_SESSION['menuItems'] = $_SESSION['menuItems'] ?? $appMenuService->getMenu();
 ?>
